rename `lib/account-password-data` to `lib/generate-password`
