+++ client/lib/generate-password/index.js
